Game Update 2 - GUI
Added a temporary and very minimal GUI to the game. A known "bug" is that you can't access a lot of the options available while in it, though this is hardly a bug and moreso just my laziness in putting out this preliminary version. Also identified a bug upon choosing users where pressing enter chooses the first user. This would actually matter if users had a difference, but it'll be looked at and fixed nonetheless before adding more features.