Game Update 2.5
Fixed the error where pressing Enter would select the last user played (or the first one, as I only had one.) Also added the ability to exit from the user selection screen back to the main menu. Expanded upon the GUI and completely changed how it was handled (previously used the pack manager; now uses the grid manager.) This greatly minimized the amount of space needed for the GUI and made the three (for now) buttons: feed, play, and clear screen. The next feature I think I will implement will be to have a button that appears when the pet is sick.